Change type of score float -> double

+++ eugene/src/GeneFeature.cc
@@ -181,7 +181,7 @@ string GeneFeature::getType()
   return type_;
 }
 
-float GeneFeature::getScore()
+double GeneFeature::getScore()
 {
   return score_;
 }
